clean up requestProps.res after it has been used
the new feature that warns about unused properties on the request
and response objects was introduced, without the res object being
removed from the request object.

that resulted in a problem where you would get "Property "res" does not
exist on the request object." errors when you tried to set
e.g. res.locals on the request object - which is a feature.