Strictly validate the response properties.
This fixes the following two corner cases of response property values:
* supplying an array
* supplying an an object with arbitrary properties

The latter is a tricky case, because unlike requests specifying an object
here does not use that as the body of the response. Instead, arbitrarily
named properties are attached to the response object. This asymmetry with
the request spec is unfortunate because with a cursory glance it can be
hard to tell a body assertion is not being made given the precedent.

This is addressed here by forcing any *non-empty* request spec to contain
at minimum a single additional property of the request and it is intended
an error in this case will better guide the user to identify the problem.